Add logic to function
We had to add parameters for the string input and the integer which
represented the number of times the user wanted their input string to
repeat.
We also had to a for loop that would build through concatenation the
return string.
And we included a condition that returned an error message when their
input integer was less than 0, because we cannot repeat a negative
number of times.