Add functional logic and parameter
This functions job is to take in a string and to reverse it.
I got a little too deep with this function and instead of reversing the
entire string argument I reversed the words in the string and then each
letter of each word, built it back into a string and returned it.

And that was way too much work because it was a simple reverse the
argument string, so I learned that I need to understand the problem a
little bit better and think about simpler approaches. Oh well.