Correct parameter usage, add number constructor
Realized that I was using the wrong parameter in the wrong function, not
sure how that happened but easy change to correct it.
When using the toFixed method, it returns a string, which I forgot, and
to change its output back to an integer we can either preface the
statement with a + operator or wrap it inside the Number constructor.