Change conditional expression to correctly filter
I figured out an error of mine was that I was trying to use typeof to
allow numbers through to the resulting arrays, however when the numbers
were being passed to typeof they were as strings and not numbers, and
I then realized typeof can't be used how I wanted for anything because
they'll all be within strings. Oops, well now I've seen that problem,
and hopefully will not do that again.
I have switched to using character codes, and have included the codes
for integers from 0 up to and including 9, so that those will be
included in the strings.
I also changed the structure of the filter method, as I believe it was
passing the outcome of the first condition and not the joint outcome
of the first and second conditions into an if elif else chain and it now
 works correctly.